break apart t and generic t test
